implementando os eventos para simulação de cloud

+++ src/ispd/gui/iconico/grade/VirtualMachine.java
@@ -57,7 +57,7 @@ public class VirtualMachine {
         this.VMM = VMM;
     }
 
-    public double getPoderComputacional() {
+    public int getPoderComputacional() {
         return numProcessadores;
     }
